refactor(backend): extract document course analysis helpers
